Finished up refactoring and prettying things up
