+++ src/components/output/Output.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { getStock } from "../util/StockUtil";
+import { getStock } from "../../util/StockUtil";
 
 export default class Output extends React.Component {
   /* utilities */
@@ -50,34 +50,36 @@ export default class Output extends React.Component {
     const stockPerformance = parseFloat((endDateValue - startDateValue).toFixed(2));
     const hasPositivePerformance = endDateValue > startDateValue;
 
-    // when parsing the date, would somehow be one day behind? Javascript is weird.
+    // when parsing the date, would somehow be one day behind? Javascript is weird
     const formattedStartDate = this.addDays(new Date(startDate), 1).toLocaleDateString();
     const formattedEndDate = this.addDays(new Date(endDate), 1).toLocaleDateString();
 
     return (
-      <div style={{ backgroundColor: "rgb(255,0,0)", display: "flex", justifyContent: "center" }}>
-        <div style={{ backgroundColor: "rgb(255,50,0)", margin: "10px 10px 10px 10px", display: "grid" }}>
-          <label style={{ fontSize: "25px" }}>
+      <div
+        style={{
+          backgroundColor: "rgb(45,45,45)",
+          textAlign: "center",
+          margin: "10px 200px 20px 200px",
+          borderRadius: "10px",
+        }}
+      >
+        <h2 style={{ fontSize: "50px", margin: "0px" }}>
+          Your Performance
+        </h2>
+        <div style={{ margin: "10px 10px 10px 10px" }}>
+          <label style={{ fontSize: "30px" }}>
             Worth in {formattedStartDate}: ${startDateValue}
-          </label>
-          <label  style={{ fontSize: "25px" }}>
+            &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
             Worth in {formattedEndDate}: ${endDateValue}
           </label>
         </div>
-        <div style={{
-          backgroundColor: "rgb(255,50,0)",
-          margin: "10px 10px 10px 10px",
-          alignItems: "center",
-          display: "flex"
-        }}>
-        <label style= {{ fontSize: "25px" }}>
+        <label style= {{ fontSize: "30px" }}>
           Performance:&nbsp;
           <span
             style= {{ color: hasPositivePerformance ? "green" : "red"}}>
             { `${hasPositivePerformance ? "+" : "-"}$${stockPerformance}` }
           </span>
         </label>
-        </div>
       </div>
     );
   }
